Draft - Introduce Interner
* Changes:

- Introduce Interner.
- Introduce boilerplate files.

* Motivation:

String interning is a fundamental technique in compilers, allowing much
more compact representations of the myriad identifiers (and literals),
and ensuring quick checks.

Unfortunately, while a quick and dirty string interner is very easily
implemented -- just a vector + map, after all -- an actual
well-performing interner is a whole other story. Most specifically, a
string interner suffers from having to assign a unique ID to all equal
strings _across threads_, which instantly makes it a prime spot for
contention.

The intent of endor being to provide solid building bricks for
compilers, it's only natural that endor would start at the very
beginning -- lexing -- and bring in a string interner which scales
near-linearly in the number of threads using it. Safely.