Introduce Benchmarks
* Changes:

- Introduce dev-dependencies on criterion & fxhash.
- Introduce benchmarks.

* Motivation:

The benchmarks establish a performance horizon (FxHashSet) and thus set
the bar for single-threaded and multi-threaded scenarios, then measure
the actual performance of Interner against this bar.

The performance of single-threaded is not too bad, taking into account
that Interner performs more work than FxHashSet: copying the string,
memorizing the length, creating a mapping, with a dose of atomic
operations.

The degradation of performance for the cooperative mode -- the idiomatic
usecase -- is a tad annoying. The "all" case is not too bad, exhibiting
similar performance to the "replicas" case, and keeping within 3x of the
horizon, but the "unique" case degenerates quite a bit, even though
there should be no true contention, which is surprising, and warrants
investigation.